// theme/theme.types.ts
export interface ITheme {
  background: string;
  textColor: string;
  headerColor: string;
  dangerColor: string;
  disabledColor: string;
  headerText: string;
  cardColor: ICard;
}

export interface ICard {
  pz: string
  lk: string
  lb: string
  zal: string
  ekz: string
  sem: string
  dop: string
}